+++ application/ImageButton.cpp
@@ -8,6 +8,8 @@
  */
 #include "ImageButton.h"
 
+#include <stdint.h>
+
 #include <algorithm>
 
 #include <Application.h>
@@ -248,12 +250,12 @@ ImageButton::_RescaleBitmap(const BBitmap* src, int32 width, int32 height)
 	void* srcData = src->Bits();
 
 	for (int32 y = 0; y <= height; y++) {
-		void* dstRow = (void*)((uint32)dstData + (uint32)(y * dstYOff));
-		void* srcRow = (void*)((uint32)srcData + ((uint32)(y * dy) * srcYOff));
+		void* dstRow = (void*)((uintptr_t)dstData + (uintptr_t)(y * dstYOff));
+		void* srcRow = (void*)((uintptr_t)srcData + ((uintptr_t)(y * dy) * srcYOff));
 
 		for (int32 x = 0; x <= width; x++)
-			memcpy((void*)((uint32)dstRow + (x * bpp)), (void*)((uint32)srcRow
-                                              + ((uint32)(x * dx) * bpp)), bpp);
+			memcpy((void*)((uintptr_t)dstRow + (x * bpp)), (void*)((uintptr_t)srcRow
+                                              + ((uintptr_t)(x * dx) * bpp)), bpp);
 	}
 
 	return res;
